+++ src/app/data/db.service.ts
@@ -161,11 +161,15 @@ export class DbService {
     return undefined;
   }
 
+  tzError = false;
   public getTimeZone(): string {
     let timezone = this.selectedDataset().timeZone;
     if (!timezone) {
       timezone = BurningManTimeZone;
-      console.error(`Shouldnt get an empty timezone`);
+      if (!this.tzError) {
+        console.error(`Shouldnt get an empty timezone`);
+        this.tzError = true;
+      }
     }
     return timezone;
   }
@@ -324,13 +328,13 @@ export class DbService {
           onlyRead = true;
         }
       }
-      
+
 
       if (!status.connected) {
         onlyRead = true;
       }
 
-      if (environment.offline) {        
+      if (environment.offline) {
         options.onlyFresh = true;
         onlyRead = false;
       }