+++ changelog.md
@@ -5,6 +5,7 @@
 - Camp image shown when pressing a pin on the map for an event
 - Fix extra line showing under reminders
 - Add friend now available from the camp menu
+- Fix crash when clicking location of camps
 
 ## 2.97
 - Fix for man pavillion location