+++ changelog.md
@@ -1,6 +1,7 @@
 ## Changes
 
 ## 2.98
+- Show friends on the map
 - Camp image is now shown on the popover for event locations
 - Camp image shown when pressing a pin on the map for an event
 - Fix extra line showing under reminders